backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
email-validator==2.1.0.post1
starlette==0.36.3

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.25.0

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# OpenRouter API integration
httpx==0.27.0
openai==1.12.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
jinja2==3.1.3

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
httpx==0.27.0